Fix extra module dependencies in persistent cache
.bas and .deps modules depended on every previously exported .bas module
due to reusing its string values for .mlb paths / ids. Simply copying
those strings fixes the issue and allows for .bas and .deps modules to be
exported without dependencies.

There is a similar dependency between .bas and .ns modules (all file
paths and identifiers used in bindings), but it does not matter: since
we only attempt to load an .ns module if its basis is still clean, that
means its .bas counterpart was already loaded in order to be checked
during dag traversal.